Website-In-Progress
The beginning files to the python charmers interactive Eliza website. Still needs a lot of editing.
To work correctly, probably need to install pip & flask before using these files. 
(That folder was too large to upload along with it).
-Using the YouTube tutorial "Python Website Full Tutorial - Flask, Authentication...." by Tech With Tim.

Note: Kind of scratched some of the text I had beforehand so we can have the option for more than one page. Still working! 
   Sorry it's taking me so long, I'm slow at learning this. Let me know if there's any problems, I'll update these files every so often.